Refactor file paths and import statements
